+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X03_ActorRequestProcessExample/V1/ActorRequestProcessExampleNotifyEventsV1.cs
@@ -12,12 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
+namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X03_ActorRequestProcessExample.V1;
 
-namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X03_ActorRequestsProcess.V1;
-
-/// <summary>
-/// Input to BRS-X03
-/// </summary>
-/// <param name="BusinessReason"></param>
-public record ActorRequestProcessExampleInputV1(string BusinessReason) : IInputParameterDto;
+public class ActorRequestProcessExampleNotifyEventsV1
+{
+    public const string ActorMessagesEnqueued = "ActorMessagesEnqueued";
+}
