+++ source/ProcessManager.Example.Orchestrations.Abstrations/Processes/BRS_X03_ActorRequestProcessExample/V1/StartActorRequestProcessExampleV1.cs
@@ -15,7 +15,7 @@
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model;
 using Energinet.DataHub.ProcessManager.Abstractions.Api.Model.OrchestrationInstance;
 
-namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X03_ActorRequestsProcess.V1;
+namespace Energinet.DataHub.ProcessManager.Example.Orchestrations.Abstractions.Processes.BRS_X03_ActorRequestProcessExample.V1;
 
 /// <summary>
 /// Start a BRS-X03 actor request process example
